chore(core): 🔧 move react-hook-form isHTMLElement tests to __tests__
